Create tabular dataset
This creates a new abstract dataset called tabular. It contains common
functionality of features and label management that is shared between the
CsvDataset, the TablesawDataset, and likely future datasets from other tabular
sources.
