+++ extensions/tablesaw/src/main/java/ai/djl/tablesaw/TablesawDataset.java
@@ -12,46 +12,43 @@
  */
 package ai.djl.tablesaw;
 
+import ai.djl.basicdataset.tabular.TabularDataset;
 import ai.djl.basicdataset.utils.DynamicBuffer;
 import ai.djl.basicdataset.utils.Feature;
 import ai.djl.ndarray.NDList;
 import ai.djl.ndarray.NDManager;
 import ai.djl.ndarray.types.Shape;
-import ai.djl.training.dataset.RandomAccessDataset;
-import ai.djl.training.dataset.Record;
 import ai.djl.util.Progress;
 import java.nio.FloatBuffer;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import tech.tablesaw.api.Row;
 import tech.tablesaw.api.Table;
 import tech.tablesaw.io.ReadOptions;
 
 /** {@code TablesawDataset} represents the dataset that stored in a .csv file. */
-public class TablesawDataset extends RandomAccessDataset {
+public class TablesawDataset extends TabularDataset {
 
     protected ReadOptions readOptions;
-    protected List<Feature> features;
-    protected List<Feature> labels;
     protected Table table;
 
     protected TablesawDataset(TablesawBuilder<?> builder) {
         super(builder);
         readOptions = builder.readOptions;
-        features = builder.features;
-        labels = builder.labels;
     }
 
     /** {@inheritDoc} */
     @Override
-    public Record get(NDManager manager, long index) {
+    public NDList getRowFeatures(NDManager manager, long index, List<Feature> selected) {
         Row row = table.row(Math.toIntExact(index));
-        NDList data = toNDList(manager, row, features);
-        NDList label = toNDList(manager, row, labels);
-
-        return new Record(data, label);
+        DynamicBuffer bb = new DynamicBuffer();
+        for (Feature feature : selected) {
+            String name = feature.getName();
+            String value = row.getString(name);
+            feature.getFeaturizer().featurize(bb, value);
+        }
+        FloatBuffer buf = bb.getBuffer();
+        return new NDList(manager.create(buf, new Shape(bb.getLength())));
     }
 
     /** {@inheritDoc} */
@@ -87,28 +84,11 @@ public class TablesawDataset extends RandomAccessDataset {
         return table.columnNames();
     }
 
-    protected NDList toNDList(NDManager manager, Row row, List<Feature> selected) {
-        DynamicBuffer bb = new DynamicBuffer();
-        for (Feature feature : selected) {
-            String name = feature.getName();
-            String value = row.getString(name);
-            feature.getFeaturizer().featurize(bb, value);
-        }
-        FloatBuffer buf = bb.getBuffer();
-        return new NDList(manager.create(buf, new Shape(bb.getLength())));
-    }
-
     /** Used to build a {@link TablesawDataset}. */
-    public static class TablesawBuilder<T extends TablesawBuilder<T>> extends BaseBuilder<T> {
+    public static class TablesawBuilder<T extends TablesawBuilder<T>>
+            extends TabularDataset.BaseBuilder<T> {
 
         protected ReadOptions readOptions;
-        protected List<Feature> features;
-        protected List<Feature> labels;
-
-        protected TablesawBuilder() {
-            features = new ArrayList<>();
-            labels = new ArrayList<>();
-        }
 
         /** {@inheritDoc} */
         @Override
@@ -129,110 +109,12 @@ public class TablesawDataset extends RandomAccessDataset {
         }
 
         /**
-         * Adds the features to the feature set.
-         *
-         * @param features the features
-         * @return this builder
-         */
-        public T addFeature(Feature... features) {
-            Collections.addAll(this.features, features);
-            return self();
-        }
-
-        /**
-         * Adds a numeric feature to the feature set.
-         *
-         * @param name the feature name
-         * @return this builder
-         */
-        public T addNumericFeature(String name) {
-            features.add(new Feature(name, true));
-            return self();
-        }
-
-        /**
-         * Adds a categorical feature to the feature set.
-         *
-         * @param name the feature name
-         * @return this builder
-         */
-        public T addCategoricalFeature(String name) {
-            features.add(new Feature(name, false));
-            return self();
-        }
-
-        /**
-         * Adds a categorical feature to the feature set with specified mapping.
-         *
-         * @param name the feature name
-         * @param map a map contains categorical value maps to index
-         * @param onehotEncode true if use onehot encode
-         * @return this builder
-         */
-        public T addCategoricalFeature(
-                String name, Map<String, Integer> map, boolean onehotEncode) {
-            features.add(new Feature(name, map, onehotEncode));
-            return self();
-        }
-
-        /**
-         * Adds the features to the label set.
-         *
-         * @param labels the labels
-         * @return this builder
-         */
-        public T addLabel(Feature... labels) {
-            Collections.addAll(this.labels, labels);
-            return self();
-        }
-
-        /**
-         * Adds a number feature to the label set.
-         *
-         * @param name the label name
-         * @return this builder
-         */
-        public T addNumericLabel(String name) {
-            labels.add(new Feature(name, true));
-            return self();
-        }
-
-        /**
-         * Adds a categorical feature to the label set.
-         *
-         * @param name the feature name
-         * @return this builder
-         */
-        public T addCategoricalLabel(String name) {
-            labels.add(new Feature(name, true));
-            return self();
-        }
-
-        /**
-         * Adds a categorical feature to the feature set with specified mapping.
-         *
-         * @param name the feature name
-         * @param map a map contains categorical value maps to index
-         * @param onehotEncode true if use onehot encode
-         * @return this builder
-         */
-        public T addCategoricalLabel(String name, Map<String, Integer> map, boolean onehotEncode) {
-            labels.add(new Feature(name, map, onehotEncode));
-            return self();
-        }
-
-        /**
          * Builds the new {@link TablesawDataset}.
          *
          * @return the new {@link TablesawDataset}
          */
         public TablesawDataset build() {
-            if (features.isEmpty()) {
-                throw new IllegalArgumentException("Missing features.");
-            }
-            if (labels.isEmpty()) {
-                throw new IllegalArgumentException("Missing labels.");
-            }
+            validate();
             return new TablesawDataset(this);
         }
     }
